types + linting
