+++ SharePointFramework/ProjectWebParts/src/components/ProjectInformation/EditPropertiesPanel/useFieldElements.tsx
@@ -10,7 +10,7 @@ import {
 } from '@fluentui/react'
 import React from 'react'
 import SPDataAdapter from '../../../data'
-import { ProjectInformationField } from '../ProjectInformationField'
+import { ProjectInformationField } from '../types'
 import { useModel } from './useModel'
 import { FieldDescription } from 'pp365-shared-library/lib/components'
