--- SharePointFramework/ProjectWebParts/src/components/ProjectInformation/ProjectInformationParentProject.ts
@@ -1,16 +1,0 @@
-import { IWeb } from '@pnp/sp/webs'
-
-export class ProjectInformationParentProject {
-  public title: string
-  public url: string
-  public childProjects: any[]
-  public iconName: 'ProductVariant' | 'ProductList'
-
-  constructor(spItem: Record<string, any>, public web: IWeb) {
-    this.title = spItem.Title
-    this.url = spItem.GtSiteUrl
-    this.childProjects = (JSON.parse(spItem.GtChildProjects ?? []) as any[]).map((i) => i.SPWebURL)
-    if (spItem.GtIsParentProject) this.iconName = 'ProductVariant'
-    else if (spItem.GtIsProgram) this.iconName = 'ProductList'
-  }
-}
